+++ commands/misc/alias.js
