+++ index.js
@@ -13,7 +13,7 @@ const TextToSpeech = require("./classes/TextToSpeech")
 // const VoiceCommands = require("./classes/VoiceCommands")
 const { bootClientFromAllVoiceChannels } = require("./classes/Helpers")
 const { ppResetDaily } = require("./commands/fun/pp")
-const { aocResetDaily } = require("./commands/fun/aocleaderboard")
+const { aocResetDaily } = require("./commands/misc/aocleaderboard")
 const { imposterReleaseCountdown } = require("./commands/fun/imposter")
 const express = require("express")
 const dotenv = require("dotenv")
